Use testcontainers instead of H2
### What's done:
* Added dependencies
* Converted DatabaseTest.kt
* Configured properties
