+++ save-backend/build.gradle.kts
@@ -23,10 +23,13 @@ dependencies {
     implementation("org.liquibase:liquibase-core:${Versions.liquibase}")
     implementation("org.hibernate:hibernate-core:${Versions.hibernate}")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa:${Versions.springBoot}")
-    implementation("com.h2database:h2:${Versions.h2}")
     implementation("org.jetbrains.kotlin:kotlin-reflect:${Versions.kotlin}")
     implementation("org.slf4j:slf4j-api:${Versions.slf4j}")
     implementation("ch.qos.logback:logback-core:${Versions.logback}")
+    implementation("mysql:mysql-connector-java:8.0.23")
+    testImplementation("org.testcontainers:testcontainers:1.15.2")
+    testImplementation("org.testcontainers:mysql:1.15.2")
+    testImplementation("org.testcontainers:junit-jupiter:1.15.2")
 }
 
 configureJacoco()
