+++ save-backend/src/main/resources/application-dev.properties
@@ -1,14 +1,7 @@
-spring.h2.console.enabled=true
-
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=test
-spring.datasource.password=test
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+# use hibernate to populate DB using data.sql
+spring.datasource.initialization-mode=always
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.validator.apply_to_ddl=false
-
-save.backend.h2-console-port=8081
