+++ save-backend/src/main/resources/application.properties
@@ -1,3 +1,4 @@
 server.port = 5000
 management.endpoints.web.exposure.include=health,info,prometheus
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
 spring.liquibase.enabled=false
